+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/collectionOperations.scala
@@ -25,7 +25,7 @@ import com.nvidia.spark.rapids.Arm._
 import com.nvidia.spark.rapids.ArrayIndexUtils.firstIndexAndNumElementUnchecked
 import com.nvidia.spark.rapids.BoolUtils.isAllValidTrue
 import com.nvidia.spark.rapids.RapidsPluginImplicits._
-import com.nvidia.spark.rapids.shims.{ComputeSequenceSizes, ShimExpression}
+import com.nvidia.spark.rapids.shims.{GetSequenceSize, ShimExpression}
 
 import org.apache.spark.sql.catalyst.analysis.{TypeCheckResult, TypeCoercion}
 import org.apache.spark.sql.catalyst.expressions.{ElementAt, ExpectsInputTypes, Expression, ImplicitCastInputTypes, NamedExpression, NullIntolerant, RowOrdering, Sequence, TimeZoneAwareExpression}
@@ -33,6 +33,7 @@ import org.apache.spark.sql.catalyst.util.GenericArrayData
 import org.apache.spark.sql.rapids.shims.RapidsErrorUtils
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.vectorized.ColumnarBatch
+import org.apache.spark.unsafe.array.ByteArrayMethods.MAX_ROUNDED_ARRAY_LENGTH
 
 case class GpuConcat(children: Seq[Expression]) extends GpuComplexTypeMergingExpression {
 
@@ -1363,11 +1364,52 @@ object GpuSequenceUtil {
     } // end of zero
   }
 
+  /**
+   * Compute the size of each sequence according to 'start', 'stop' and 'step'.
+   * A row (Row[start, stop, step]) contains at least one null element will produce
+   * a null in the output.
+   *
+   * The returned column should be closed.
+   */
+  def computeSequenceSize(
+      start: ColumnVector,
+      stop: ColumnVector,
+      step: ColumnVector): ColumnVector = {
+    checkSequenceInputs(start, stop, step)
+    val actualSize = GetSequenceSize(start, stop, step)
+    val sizeAsLong = withResource(actualSize) { _ =>
+      val mergedEquals = withResource(start.equalTo(stop)) { equals =>
+        if (step.hasNulls) {
+          // Also set the row to null where step is null.
+          equals.mergeAndSetValidity(BinaryOp.BITWISE_AND, equals, step)
+        } else {
+          equals.incRefCount()
+        }
+      }
+      withResource(mergedEquals) { _ =>
+        withResource(Scalar.fromLong(1L)) { one =>
+          mergedEquals.ifElse(one, actualSize)
+        }
+      }
+    }
+    withResource(sizeAsLong) { _ =>
+      // check max size
+      withResource(Scalar.fromInt(MAX_ROUNDED_ARRAY_LENGTH)) { maxLen =>
+        withResource(sizeAsLong.lessOrEqualTo(maxLen)) { allValid =>
+          require(isAllValidTrue(allValid), GetSequenceSize.TOO_LONG_SEQUENCE)
+        }
+      }
+      // cast to int and return
+      sizeAsLong.castTo(DType.INT32)
+    }
+  }
 }
 
 case class GpuSequence(start: Expression, stop: Expression, stepOpt: Option[Expression],
     timeZoneId: Option[String] = None) extends TimeZoneAwareExpression with GpuExpression
     with ShimExpression {
+
+  import GpuSequenceUtil._
 
   override def dataType: ArrayType = ArrayType(start.dataType, containsNull = false)
 
@@ -1394,7 +1436,7 @@ case class GpuSequence(start: Expression, stop: Expression, stepOpt: Option[Expr
           val steps = stepGpuColOpt.map(_.getBase.incRefCount())
               .getOrElse(defaultStepsFunc(startCol, stopCol))
           closeOnExcept(steps) { _ =>
-            (ComputeSequenceSizes(startCol, stopCol, steps), steps)
+            (computeSequenceSize(startCol, stopCol, steps), steps)
           }
         }
