+++ sql-plugin/src/main/spark311/scala/com/nvidia/spark/rapids/shims/ComputeSequenceSize.scala
@@ -42,12 +42,11 @@ package com.nvidia.spark.rapids.shims
 
 import ai.rapids.cudf._
 import com.nvidia.spark.rapids.Arm._
-import com.nvidia.spark.rapids.BoolUtils.isAllValidTrue
 
-import org.apache.spark.sql.rapids.GpuSequenceUtil
 import org.apache.spark.unsafe.array.ByteArrayMethods.MAX_ROUNDED_ARRAY_LENGTH
 
-object ComputeSequenceSizes {
+object GetSequenceSize {
+  val TOO_LONG_SEQUENCE = s"Too long sequence found. Should be <= $MAX_ROUNDED_ARRAY_LENGTH"
   /**
    * Compute the size of each sequence according to 'start', 'stop' and 'step'.
    * A row (Row[start, stop, step]) contains at least one null element will produce
@@ -59,7 +58,6 @@ object ComputeSequenceSizes {
       start: ColumnVector,
       stop: ColumnVector,
       step: ColumnVector): ColumnVector = {
-    GpuSequenceUtil.checkSequenceInputs(start, stop, step)
     // Spark's algorithm to get the length (aka size)
     // ``` Scala
     //   size = if (start == stop) 1L else 1L + (stop.toLong - start.toLong) / estimatedStep.toLong
@@ -67,46 +65,22 @@ object ComputeSequenceSizes {
     //       s"Too long sequence: $len. Should be <= $MAX_ROUNDED_ARRAY_LENGTH")
     //   size.toInt
     // ```
-    val sizeAsLong = withResource(Scalar.fromLong(1L)) { one =>
-      val diff = withResource(stop.castTo(DType.INT64)) { stopAsLong =>
-        withResource(start.castTo(DType.INT64)) { startAsLong =>
-          stopAsLong.sub(startAsLong)
-        }
+    val diff = withResource(stop.castTo(DType.INT64)) { stopAsLong =>
+      withResource(start.castTo(DType.INT64)) { startAsLong =>
+        stopAsLong.sub(startAsLong)
       }
-      val quotient = withResource(diff) { _ =>
-        withResource(step.castTo(DType.INT64)) { stepAsLong =>
-          diff.div(stepAsLong)
-        }
+    }
+    val quotient = withResource(diff) { _ =>
+      withResource(step.castTo(DType.INT64)) { stepAsLong =>
+        diff.div(stepAsLong)
       }
-      // actualSize = 1L + (stop.toLong - start.toLong) / estimatedStep.toLong
-      val actualSize = withResource(quotient) { quotient =>
+    }
+    // actualSize = 1L + (stop.toLong - start.toLong) / estimatedStep.toLong
+    val actualSize = withResource(Scalar.fromLong(1L)) { one =>
+      withResource(quotient) { quotient =>
         quotient.add(one, DType.INT64)
       }
-      withResource(actualSize) { _ =>
-        val mergedEquals = withResource(start.equalTo(stop)) { equals =>
-          if (step.hasNulls) {
-            // Also set the row to null where step is null.
-            equals.mergeAndSetValidity(BinaryOp.BITWISE_AND, equals, step)
-          } else {
-            equals.incRefCount()
-          }
-        }
-        withResource(mergedEquals) { _ =>
-          mergedEquals.ifElse(one, actualSize)
-        }
-      }
     }
-
-    withResource(sizeAsLong) { _ =>
-      // check max size
-      withResource(Scalar.fromInt(MAX_ROUNDED_ARRAY_LENGTH)) { maxLen =>
-        withResource(sizeAsLong.lessOrEqualTo(maxLen)) { allValid =>
-          require(isAllValidTrue(allValid),
-            s"Too long sequence found. Should be <= $MAX_ROUNDED_ARRAY_LENGTH")
-        }
-      }
-      // cast to int and return
-      sizeAsLong.castTo(DType.INT32)
-    }
+    actualSize
   }
 }
